[preview] Disable linkedin secrets on GITPOD_WITH_DEDICATED_EMU

+++ dev/preview/workflow/preview/deploy-gitpod.sh
@@ -382,13 +382,16 @@ fi
 #
 # configureLinkedIn
 #
-kubectl --kubeconfig "${DEV_KUBE_PATH}" --context "${DEV_KUBE_CONTEXT}" -n werft get secret linked-in -o yaml > linked-in.secret.yaml
-yq w -i linked-in.secret.yaml metadata.namespace "default"
-yq d -i linked-in.secret.yaml metadata.creationTimestamp
-yq d -i linked-in.secret.yaml metadata.uid
-yq d -i linked-in.secret.yaml metadata.resourceVersion
-diff-apply "${PREVIEW_K3S_KUBE_CONTEXT}" linked-in.secret.yaml
-rm -f linked-in.secret.yaml
+if [[ "${GITPOD_WITH_DEDICATED_EMU}" != "true" ]]
+then
+  kubectl --kubeconfig "${DEV_KUBE_PATH}" --context "${DEV_KUBE_CONTEXT}" -n werft get secret linked-in -o yaml > linked-in.secret.yaml
+  yq w -i linked-in.secret.yaml metadata.namespace "default"
+  yq d -i linked-in.secret.yaml metadata.creationTimestamp
+  yq d -i linked-in.secret.yaml metadata.uid
+  yq d -i linked-in.secret.yaml metadata.resourceVersion
+  diff-apply "${PREVIEW_K3S_KUBE_CONTEXT}" linked-in.secret.yaml
+  rm -f linked-in.secret.yaml
+fi
 
 #
 # configureSSHGateway
